# nooblyjs-application-wiki/backend/swagger/openapi.full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NooblyJS Wiki Platform API",
    "description": "Unified REST API for the NooblyJS Wiki Platform (port 9101). Covers the Datasources module (`/api/*`), Wiki module (`/applications/wiki/api/*`) and Blueprint module (`/applications/blueprint/api/*`).",
    "version": "2.0.0",
    "contact": { "name": "NooblyJS Team" },
    "license": { "name": "ISC" }
  },
  "servers": [
    { "url": "/", "description": "Unified backend (port 9101)" }
  ],
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie issued after login"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Bearer token (24 hr). Obtain from POST /api/auth/token or the API tokens section of your profile."
      }
    },
...
  },
  "security": [
    { "SessionAuth": [] },
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/security/user": {
      "post": {
        "tags": ["Datasources - Security"],
        "summary": "Add a user to a role (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": { "type": "string" }, "role": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Added" } }
      },
      "delete": {
        "tags": ["Datasources - Security"],
        "summary": "Remove a user from a role (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": { "type": "string" }, "role": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Removed" } }
      }
    },
    "/api/security/admin": {
      "post": {
        "tags": ["Datasources - Security"],
        "summary": "Promote a user to admin (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Promoted" } }
      },
      "delete": {
        "tags": ["Datasources - Security"],
        "summary": "Revoke admin from a user (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Revoked" } }
      }
    },
...
  }
}